// projects/anyopsos-app-monitor/src/lib/types/netdata.ts
import {VMWareObject} from '@anyopsos/app-infrastructure-manager';

export interface Netdata {
  uuid: string;
  url: string;
  description: string;
  credential: string;
  withCredential: boolean;
  linkTo: VMWareObject;
  autologin: boolean;
  save: boolean;
  state: string;
  type: 'netdata' | 'netdata-credential' | 'snapshot' | 'internal';
  error?: string;
  snapshotData?: any;
}